backend/requirements.txt
fastapi>=0.115.0
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic-settings==2.0.3
python-multipart>=0.0.18
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
passlib
bcrypt
python-dotenv
pydantic-settings
python-multipart
pydantic[email]
